requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
pyyaml
matplotlib
tqdm
h5py
gdown
